Fix bugs and create tests

+++ app/src/main/java/com/example/android/bakingapp/Ingredient.java
@@ -16,7 +16,7 @@ public class Ingredient implements Parcelable{
     }
 
     protected Ingredient(Parcel in) {
-        quantity = in.readInt();
+        quantity = in.readDouble();
         measure = in.readString();
         name = in.readString();
     }
